Fix some implicit casting in ASP
I'm trying to stomp down some of the -Wconversion warnings that I've
been finding in VW.

+++ src/asp/Sessions/ISIS/StereoSessionIsis.h
@@ -102,7 +102,7 @@ public:
   PixelT operator() (PixelT const& pix) const {
     typedef typename CompoundChannelType<PixelT>::type channel_type;
     typedef Helper<channel_type, void> help;
-    for (int n = 0; n < CompoundNumChannels<PixelT>::value; ++n) {
+    for (size_t n = 0; n < CompoundNumChannels<PixelT>::value; ++n) {
       if (help::IsSpecial(compound_select_channel<const channel_type&>(pix,n)))
         if (help::IsHighPixel(compound_select_channel<const channel_type&>(pix,n)))
           return m_replacement_high;
